fizz-cli/pyproject.toml
[tool.poetry]
name = "fizz-cli"
version = "0.1.0"
description = ""
authors = ["Kaustubh Kishore <KaustubhKishorem@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
typer = {extras = ["all"], version = "^0.9.0"}
bullet = {version="^2.2.0"}
PyYAML = {version="^6.0.1"}

[tool.poetry.scripts]
fizz = "fizz_cli.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 130
string-normalization = true
force-alphabetical-sort-within-sections = true